# Fix `Key` interning race.