@jsirois 2021-05-31
Although this fix is needed for correctness, I had never measured the perf impact. #11262 noted a ~5% speedup for `./pants dependencies --transitive ::` with interning implemented.

This change looks like its affect on that benchmark is in the noise. Here `~/dev/pantsbuild/pants` is master and `~/dev/pantsbuild/jsirois-pants` is this PR:

With pantsd:
```
$ hyperfine -w 2 -L 'clone' 'pants,jsirois-pants' 'cd ~/dev/pantsbuild/{clone} && ./pants dependencies --transitive ::'
Benchmark #1: cd ~/dev/pantsbuild/pants && ./pants dependencies --transitive ::
  Time (mean ± σ):      3.012 s ±  0.055 s    [User: 1.048 s, System: 0.047 s]
  Range (min … max):    2.945 s …  3.100 s    10 runs
 
Benchmark #2: cd ~/dev/pantsbuild/jsirois-pants && ./pants dependencies --transitive ::
  Time (mean ± σ):      2.976 s ±  0.050 s    [User: 1.033 s, System: 0.053 s]
  Range (min … max):    2.907 s …  3.051 s    10 runs
 
Summary
  'cd ~/dev/pantsbuild/jsirois-pants && ./pants dependencies --transitive ::' ran
    1.01 ± 0.03 times faster than 'cd ~/dev/pantsbuild/pants && ./pants dependencies --transitive ::'
```

Without pantsd:
```
$ hyperfine -w 2 -L 'clone' 'pants,jsirois-pants' 'cd ~/dev/pantsbuild/{clone} && ./pants --no-pantsd dependencies --transitive ::'
Benchmark #1: cd ~/dev/pantsbuild/pants && ./pants --no-pantsd dependencies --transitive ::
  Time (mean ± σ):     10.650 s ±  0.085 s    [User: 12.206 s, System: 3.347 s]
  Range (min … max):   10.547 s … 10.832 s    10 runs
 
Benchmark #2: cd ~/dev/pantsbuild/jsirois-pants && ./pants --no-pantsd dependencies --transitive ::
  Time (mean ± σ):     10.605 s ±  0.181 s    [User: 12.170 s, System: 3.287 s]
  Range (min … max):   10.451 s … 10.944 s    10 runs
 
Summary
  'cd ~/dev/pantsbuild/jsirois-pants && ./pants --no-pantsd dependencies --transitive ::' ran
    1.00 ± 0.02 times faster than 'cd ~/dev/pantsbuild/pants && ./pants --no-pantsd dependencies --transitive ::'
```