@jsirois 2021-06-01
> Could you please cherry-pick to 2.4 and 2.5? I will get out release candidates with this.

Will do. You cut 2.3.2 and 2.2.3 on May 7th and this bug goes back to 2.2.x. That implies we're treating 2.2.x and greater as maintained still and so this cherrypick should got to those two as well. I'll optimistically prepare those and you can shoot down if there is something I'm missing.