@Eric-Arellano 2021-06-01
Sure, 2.3 and 2.2 sound great as well. Thanks