@jsirois 2021-06-02
Now cherry-picked to `2.{2,3,4,5}.x`.